+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 ## 0.65.0 [Master]
 
 `tonal-note`
-- TODO: add oct
+- Expose note properties: props, oct, alt, step
 
 `tonal-array`
 - TODO: sort: change order of parameters to make it more JS friendly
+
+`tonal-pitchset`
+- TODO: rename toBinary to chroma
 
 ## 0.64.1